added version checks to all install functions

+++ shell/common/nredf_functions.sh
@@ -70,6 +70,9 @@ function _nredf_set_defaults() {
 function _nredf_install_fzf() {
   _nredf_get_sys_info
   local VERSION=$(_nredf_github_latest_release junegunn fzf)
+
+  [[ "${VERSION}" == "" ]] && return 1
+
   if [[ ! -f "${HOME}/.local/bin/fzf" ]] || [[ "${VERSION}" != "$(${HOME}/.local/bin/fzf --version | awk '{print $1}')" ]]; then
     echo -e '\033[1mInstalling fzf\033[0m'
     [[ -d ${HOME}/.fzf ]] && rm -rf ${HOME}/.fzf
@@ -87,8 +90,12 @@ function _nredf_install_fzf() {
 
 function _nredf_install_nvim() {
   _nredf_get_sys_info
+  local VERSION=$(curl -sH 'Accept: application/vnd.github.v3+json' https://api.github.com/repos/neovim/neovim/releases/tags/nightly | grep -Po '"name":"\K.*?(?=")' | head -1)
+  local SHELL_NAME=$(readlink /proc/$$/exe | awk -F'/' '{print $NF}')
 
+  [[ "${VERSION}" == "" ]] && return 1
   [[ "${OS}" != "linux" ]] && return 1
+
   if [[ ! -f "${HOME}/.local/bin/nvim" ]] || [[ "$(${HOME}/.local/bin/nvim --version | head -1)" != "$(curl -sH 'Accept: application/vnd.github.v3+json' https://api.github.com/repos/neovim/neovim/releases/tags/nightly | grep -Po '"name":"\K.*?(?=")' | head -1)" ]]; then
     echo -e '\033[1mDownloading neovim\033[0m'
     [[ -d "${HOME}/.cache/vim/squashfs-root" ]] && rm -rf "${HOME}/.cache/vim/squashfs-root"
@@ -107,71 +114,134 @@ function _nredf_install_nvim() {
 function _nredf_install_lf() {
   _nredf_get_sys_info
 
-  export LF_ICONS="tw=:st=:ow=:dt=:di=:fi=:ln=:or=:ex=:*.c=:*.cc=:*.clj=:*.coffee=:*.cpp=:*.css=:*.d=:*.dart=:*.erl=:*.exs=:*.fs=:*.go=:*.h=:*.hh=:*.hpp=:*.hs=:*.html=:*.java=:*.jl=:*.js=:*.json=:*.lua=:*.md=:*.php=:*.pl=:*.pro=:*.py=:*.rb=:*.rs=:*.scala=:*.ts=:*.vim=:*.cmd=:*.ps1=:*.sh=:*.bash=:*.zsh=:*.fish=:*.tar=:*.tgz=:*.arc=:*.arj=:*.taz=:*.lha=:*.lz4=:*.lzh=:*.lzma=:*.tlz=:*.txz=:*.tzo=:*.t7z=:*.zip=:*.z=:*.dz=:*.gz=:*.lrz=:*.lz=:*.lzo=:*.xz=:*.zst=:*.tzst=:*.bz2=:*.bz=:*.tbz=:*.tbz2=:*.tz=:*.deb=:*.rpm=:*.jar=:*.war=:*.ear=:*.sar=:*.rar=:*.alz=:*.ace=:*.zoo=:*.cpio=:*.7z=:*.rz=:*.cab=:*.wim=:*.swm=:*.dwm=:*.esd=:*.jpg=:*.jpeg=:*.mjpg=:*.mjpeg=:*.gif=:*.bmp=:*.pbm=:*.pgm=:*.ppm=:*.tga=:*.xbm=:*.xpm=:*.tif=:*.tiff=:*.png=:*.svg=:*.svgz=:*.mng=:*.pcx=:*.mov=:*.mpg=:*.mpeg=:*.m2v=:*.mkv=:*.webm=:*.ogm=:*.mp4=:*.m4v=:*.mp4v=:*.vob=:*.qt=:*.nuv=:*.wmv=:*.asf=:*.rm=:*.rmvb=:*.flc=:*.avi=:*.fli=:*.flv=:*.gl=:*.dl=:*.xcf=:*.xwd=:*.yuv=:*.cgm=:*.emf=:*.ogv=:*.ogx=:*.aac=:*.au=:*.flac=:*.m4a=:*.mid=:*.midi=:*.mka=:*.mp3=:*.mpc=:*.ogg=:*.ra=:*.wav=:*.oga=:*.opus=:*.spx=:*.xspf=:*.pdf=:*.nix=:"
-  if [[ ! -f "${HOME}/.local/bin/lf" ]] || [[ "$(_nredf_github_latest_release gokcehan lf)" != "$(${HOME}/.local/bin/lf -version)" ]]; then
+  local VERSION=$(_nredf_github_latest_release gokcehan lf)
+  [[ "${VERSION}" == "" ]] && return 1
+
+  if [[ ! -f "${HOME}/.local/bin/lf" ]] || [[ "${VERSION}" != "$(${HOME}/.local/bin/lf -version)" ]]; then
     echo -e '\033[1mInstalling lf\033[0m'
     curl -Ls "https://github.com/gokcehan/lf/releases/latest/download/lf-${OS}-${ARCH}.tar.gz" | tar xzf - -C ${HOME}/.local/bin/
     chmod +x ${HOME}/.local/bin/lf
+    export LF_ICONS="tw=:st=:ow=:dt=:di=:fi=:ln=:or=:ex=:*.c=:*.cc=:*.clj=:*.coffee=:*.cpp=:*.css=:*.d=:*.dart=:*.erl=:*.exs=:*.fs=:*.go=:*.h=:*.hh=:*.hpp=:*.hs=:*.html=:*.java=:*.jl=:*.js=:*.json=:*.lua=:*.md=:*.php=:*.pl=:*.pro=:*.py=:*.rb=:*.rs=:*.scala=:*.ts=:*.vim=:*.cmd=:*.ps1=:*.sh=:*.bash=:*.zsh=:*.fish=:*.tar=:*.tgz=:*.arc=:*.arj=:*.taz=:*.lha=:*.lz4=:*.lzh=:*.lzma=:*.tlz=:*.txz=:*.tzo=:*.t7z=:*.zip=:*.z=:*.dz=:*.gz=:*.lrz=:*.lz=:*.lzo=:*.xz=:*.zst=:*.tzst=:*.bz2=:*.bz=:*.tbz=:*.tbz2=:*.tz=:*.deb=:*.rpm=:*.jar=:*.war=:*.ear=:*.sar=:*.rar=:*.alz=:*.ace=:*.zoo=:*.cpio=:*.7z=:*.rz=:*.cab=:*.wim=:*.swm=:*.dwm=:*.esd=:*.jpg=:*.jpeg=:*.mjpg=:*.mjpeg=:*.gif=:*.bmp=:*.pbm=:*.pgm=:*.ppm=:*.tga=:*.xbm=:*.xpm=:*.tif=:*.tiff=:*.png=:*.svg=:*.svgz=:*.mng=:*.pcx=:*.mov=:*.mpg=:*.mpeg=:*.m2v=:*.mkv=:*.webm=:*.ogm=:*.mp4=:*.m4v=:*.mp4v=:*.vob=:*.qt=:*.nuv=:*.wmv=:*.asf=:*.rm=:*.rmvb=:*.flc=:*.avi=:*.fli=:*.flv=:*.gl=:*.dl=:*.xcf=:*.xwd=:*.yuv=:*.cgm=:*.emf=:*.ogv=:*.ogx=:*.aac=:*.au=:*.flac=:*.m4a=:*.mid=:*.midi=:*.mka=:*.mp3=:*.mpc=:*.ogg=:*.ra=:*.wav=:*.oga=:*.opus=:*.spx=:*.xspf=:*.pdf=:*.nix=:"
   fi
 }
 
 function _nredf_install_k8s_ops() {
-  _nredf_get_sys_info
+  _nredf_install_kubectl
+  _nredf_install_krew
+  _nredf_install_kubeadm
+  _nredf_install_fluxcd
+  _nredf_install_helm
+  _nredf_install_k9s
+}
 
-  if [[ ! -f "${HOME}/.local/bin/kubectl" ]] || [[ $(curl -L -s https://dl.k8s.io/release/stable.txt) != $(${HOME}/.local/bin/kubectl version --short --client | awk -F: '{ gsub(/ /,""); print $2}') ]]; then
+function _nredf_install_kubectl() {
+  _nredf_get_sys_info
+  local VERSION=$(curl -L -s https://dl.k8s.io/release/stable.txt)
+  local SHELL_NAME=$(readlink /proc/$$/exe | awk -F'/' '{print $NF}')
+
+  [[ "${VERSION}" == "" ]] && return 1
+
+  if [[ ! -f "${HOME}/.local/bin/kubectl" ]] || [[ ${VERSION} != $(${HOME}/.local/bin/kubectl version --short --client | awk -F: '{ gsub(/ /,""); print $2}') ]]; then
     echo -e '\033[1mInstalling kubectl\033[0m'
-    curl -Ls "https://dl.k8s.io/release/$(curl -L -s https://dl.k8s.io/release/stable.txt)/bin/${OS}/${ARCH}/kubectl" -o ${HOME}/.local/bin/kubectl
+    curl -Ls "https://dl.k8s.io/release/${VERSION}/bin/${OS}/${ARCH}/kubectl" -o ${HOME}/.local/bin/kubectl
     chmod +x ${HOME}/.local/bin/kubectl
   fi
 
-  if [[ ! -f "${HOME}/.local/bin/kubeadm" ]] || [[ $(curl -L -s https://dl.k8s.io/release/stable.txt) != $(${HOME}/.local/bin/kubeadm version --short --client | awk -F: '{ gsub(/ /,""); print $2}') ]]; then
-    echo -e '\033[1mInstalling kubeadm\033[0m'
-    curl -Ls "https://dl.k8s.io/release/$(curl -L -s https://dl.k8s.io/release/stable.txt)/bin/${OS}/${ARCH}/kubectl" -o ${HOME}/.local/bin/kubeadm
-    chmod +x ${HOME}/.local/bin/kubeadm
+  if [[ "${SHELL_NAME}" =~ "bash|zsh" ]]; then
+    [[ -f ${HOME}/.local/bin/kubectl ]] && source <(${HOME}/.local/bin/kubectl completion ${SHELL_NAME})
   fi
+}
 
-  if [[ ! -d "${HOME}/.krew" ]]; then
+function _nredf_install_krew() {
+  [[ ! -f "${HOME}/.local/bin/kubectl" ]] && return 1
+
+  local VERSION=$(_nredf_github_latest_release gokcehan lf)
+
+  [[ "${VERSION}" == "" ]] && return 1
+
+  if [[ ! -d "${HOME}/.krew" ]] || [[ ${VERSION} != $(${HOME}/.local/bin/kubectl krew version | awk '/^GitTag/{print $2}') ]]; then
     echo -e '\033[1mInstalling krew\033[0m'
     curl -fsSLo ${HOME}/.cache/krew/krew.tar.gz "https://github.com/kubernetes-sigs/krew/releases/latest/download/krew.tar.gz"
     tar -zxf ${HOME}/.cache/krew/krew.tar.gz --directory ${HOME}/.cache/krew && rm -f ${HOME}/.cache/krew/krew.tar.gz
     KREW=${HOME}/.cache/krew/krew-"$(uname | tr '[:upper:]' '[:lower:]')_$(uname -m | sed -e 's/x86_64/amd64/' -e 's/arm.*$/arm/')"
     "$KREW" install krew 2>/dev/null
     export PATH="${KREW_ROOT:-$HOME/.krew}/bin:$PATH"
-  else
-    echo -e '\033[1mUpdating krew and plugins\033[0m'
-    export PATH="${KREW_ROOT:-$HOME/.krew}/bin:$PATH"
-    kubectl krew update 2>/dev/null
-    kubectl krew upgrade 2>/dev/null
-    for KREW_PLUGIN in ctx ns doctor fuzzy images status oidc-login; do
-      kubectl krew list | grep -q ${KREW_PLUGIN} || kubectl krew install ${KREW_PLUGIN} 2>/dev/null
-    done
   fi
 
-  if [[ ! -f "${HOME}/.local/bin/fluxctl" ]] || [[ ! "$(_nredf_github_latest_release fluxcd flux)" == "$(${HOME}/.local/bin/fluxctl version)" ]]; then
+  echo -e '\033[1mUpdating krew plugins\033[0m'
+  export PATH="${KREW_ROOT:-$HOME/.krew}/bin:$PATH"
+  kubectl krew upgrade 2>/dev/null
+  for KREW_PLUGIN in ctx ns doctor fuzzy images status oidc-login; do
+    kubectl krew list | grep -q ${KREW_PLUGIN} || kubectl krew install ${KREW_PLUGIN} 2>/dev/null
+  done
+}
+
+function _nredf_install_kubeadm() {
+  _nredf_get_sys_info
+  local VERSION=$(curl -L -s https://dl.k8s.io/release/stable.txt)
+  local SHELL_NAME=$(readlink /proc/$$/exe | awk -F'/' '{print $NF}')
+
+  [[ "${VERSION}" == "" ]] && return 1
+
+  if [[ ! -f "${HOME}/.local/bin/kubeadm" ]] || [[ ${VERSION} != $(${HOME}/.local/bin/kubeadm version --short --client | awk -F: '{ gsub(/ /,""); print $2}') ]]; then
+    echo -e '\033[1mInstalling kubeadm\033[0m'
+    curl -Ls "https://dl.k8s.io/release/${VERSION}/bin/${OS}/${ARCH}/kubectl" -o ${HOME}/.local/bin/kubeadm
+    chmod +x ${HOME}/.local/bin/kubeadm
+  fi
+
+  if [[ "${SHELL_NAME}" =~ "bash|zsh" ]]; then
+    [[ -f ${HOME}/.local/bin/kubeadm ]] && source <(${HOME}/.local/bin/kubeadm completion ${SHELL_NAME})
+  fi
+}
+
+function _nredf_install_fluxcd() {
+  _nredf_get_sys_info
+  local VERSION=$(_nredf_github_latest_release fluxcd flux)
+  local SHELL_NAME=$(readlink /proc/$$/exe | awk -F'/' '{print $NF}')
+
+  [[ "${VERSION}" == "" ]] && return 1
+
+  if [[ ! -f "${HOME}/.local/bin/fluxctl" ]] || [[ ! "${VERSION}" == "$(${HOME}/.local/bin/fluxctl version)" ]]; then
     echo -e '\033[1mInstalling fluxctl\033[0m'
     [[ -f "${HOME}/.local/bin/fluxctl" ]] && rm -f "${HOME}/.local/bin/fluxctl"
     curl -sL https://github.com/fluxcd/flux/releases/latest/download/fluxctl_${OS}_${ARCH} -o ${HOME}/.local/bin/fluxctl
     chmod +x ${HOME}/.local/bin/fluxctl
   fi
 
-  echo -e '\033[1mRunning get_helm\033[0m'
-  curl -fsSL -o ${HOME}/.cache/helm/get_helm.sh https://raw.githubusercontent.com/helm/helm/master/scripts/get-helm-3
-  chmod +x ${HOME}/.cache/helm/get_helm.sh
-  HELM_INSTALL_DIR="${HOME}/.local/bin" ${HOME}/.cache/helm/get_helm.sh --no-sudo >/dev/null
+  if [[ "${SHELL_NAME}" =~ "bash|zsh" ]]; then
+    [[ -f ${HOME}/.local/bin/fluxctl ]] && source <(${HOME}/.local/bin/fluxctl completion ${SHELL_NAME})
+  fi
+}
 
-  if [[ ! -f "${HOME}/.local/bin/k9s" ]] || [[ ! "$(_nredf_github_latest_release derailed k9s)" == "$(${HOME}/.local/bin/k9s version | grep Version | awk '{print $2}' | sed -r "s/\x1B\[([0-9]{1,2}(;[0-9]{1,2})?)?[m|K]//g")" ]]; then
+function _nredf_install_helm() {
+  _nredf_get_sys_info
+  local VERSION=$(_nredf_github_latest_release helm helm)
+  local SHELL_NAME=$(readlink /proc/$$/exe | awk -F'/' '{print $NF}')
+
+  [[ "${VERSION}" == "" ]] && return 1
+
+  if [[ ! -f "${HOME}/.local/bin/helm" ]] || [[ ! "${VERSION}" == "$(${HOME}/.local/bin/helm version --template='{{ .Version }}')" ]]; then
+    echo -e '\033[1mRunning get_helm\033[0m'
+    [[ -f "${HOME}/.local/bin/helm" ]] && rm -f "${HOME}/.local/bin/helm"
+    curl -sL https://get.helm.sh/helm-${VERSION}-${OS}-${ARCH}.tar.gz | tar xzf - -C ${HOME}/.local/bin/
+    chmod +x ${HOME}/.local/bin/helm
+  fi
+
+  if [[ "${SHELL_NAME}" =~ "bash|zsh" ]]; then
+    [[ -f ${HOME}/.local/bin/helm ]] && source <(${HOME}/.local/bin/helm completion ${SHELL_NAME})
+  fi
+}
+
+function _nredf_install_k9s() {
+  _nredf_get_sys_info
+  local VERSION=$(_nredf_github_latest_release derailed k9s)
+
+  [[ "${VERSION}" == "" ]] && return 1
+
+  if [[ ! -f "${HOME}/.local/bin/k9s" ]] || [[ ! "${VERSION}" == "$(${HOME}/.local/bin/k9s version | grep Version | awk '{print $2}' | sed -r "s/\x1B\[([0-9]{1,2}(;[0-9]{1,2})?)?[m|K]//g")" ]]; then
     echo -e '\033[1mInstalling k9s\033[0m'
     [[ -f "${HOME}/.local/bin/k9s" ]] && rm -f "${HOME}/.local/bin/k9s"
     curl -Ls "https://github.com/derailed/k9s/releases/latest/download/k9s_$(uname)_$(uname -m).tar.gz" | tar xzf - -C ${HOME}/.local/bin/ k9s
     chmod +x ${HOME}/.local/bin/k9s
   fi
-
-  local SHELL_NAME=$(readlink /proc/$$/exe | awk -F'/' '{print $NF}')
-  if [[ "${SHELL_NAME}" =~ "bash|zsh" ]]; then
-    [[ -f ${HOME}/.local/bin/kubectl ]] && source <(${HOME}/.local/bin/kubectl completion ${SHELL_NAME})
-    [[ -f ${HOME}/.local/bin/fluxctl ]] && source <(${HOME}/.local/bin/fluxctl completion ${SHELL_NAME})
-    [[ -f ${HOME}/.local/bin/helm ]] && source <(${HOME}/.local/bin/helm completion ${SHELL_NAME})
-  fi
-
 }
-
